Updated to v0.4
Fixed a bug where the current Team Score would be displayed in the wrong
order in TrackMania2.
This is due to TM2 sorting the team score by the highest points, rather
than sorting it by teams (like TMNF does).
A version check was added to the function "ovrly_show" to determine if
TMNF or TM2 is being played, because the changed part of the function
requires a TM2-only method called "GetCurrentWinnerTeam".
This should still work just fine with TMNF, since the code for TMNF
should not have been touched.